Allow to use `Server.Name` as the `Node` name of a `Machine` (#103)
* Allow to use `Server.Name` as the `Node` name of a `Machine`

This PR introduces a new flag to the MCM driver
(`use-server-name-as-node-name`) which uses the underlying `Server` name
as the `Node` name instead of relying on the `ServerClaim` name.

In this context the `CreateMachine` flow has been slightly modified as
it uses now a `wait.PollUntilContextTimeout` when creating a
`ServerClaim`. That way we can be sure, that a `Server` has been
successfully assigned to a `ServerClaim`.

* Change hostname in `ServerClaim` igintion

Use conditioanlly either the `ServerClaim` name or the claimed `Server`
name in the Ignition.

* Use `node-name-policy` typed arg instead of bool